pyproject.toml
[tool.poetry]
name = "subt-ai-tles"
version = "0.1.0"
description = ""
authors = ["xyzjonas <you@example.com>"]
license = "MIT"
readme = "README.md"

packages = [
    { from = "src", include = "subtaitles" }
]

[tool.poetry.scripts]
subtaitles = "subtaitles.cli:start_app"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.3.6"
python-dotenv = "^1.0.0"
httpx = "^0.25.2"
libretranslatepy = "^2.1.3"
pysrt = "^1.1.2"
deepl = "^1.16.1"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.0"
fastapi = "^0.104.1"
jinja2 = "^3.1.2"
uvicorn = "^0.24.0.post1"
python-multipart = "^0.0.6"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.2"
pytest-dotenv = "^0.5.2"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
